baf: fix Status_Return g_ms2pc negative-index OOB (found by fuzzing)
The 2h libFuzzer campaign immediately found a real out-of-bounds read:
Status_Return reads dwState into a signed int, so a client value with the high
bit set makes `dwState % 6` negative and indexes g_ms2pc[] (int[7]) out of
bounds. Fixed with an unsigned modulo; added a regression vector (27 checks).

The campaign's other findings were parse.h's x86 fast-path
in_uint32_le/out_uint32_le doing unaligned int access -- pre-existing,
xrdp-wide, x86-benign UB already avoided under NEED_ALIGN, not a smart-card
bug. The fuzzer and the lab runner exclude it with -fno-sanitize=alignment.

test-lab/kvm/scripts/run-smartcard-tests.sh: --fuzz-seconds now builds the
whole tree with clang + fuzzer-no-link (so libcommon and the fuzzer share one
runtime), seeds a per-selector corpus, runs fork-mode, and reports artifacts.
UPSTREAM-MS-RDPESC-REVIEW.md records both findings.

Verified: 27/27 vectors pass (ASan/UBSan); a fresh fuzz smoke is crash-free.

+++ sesman/chansrv/smartcard_pcsc.c
@@ -1637,7 +1637,9 @@ scard_function_status_return(void *user_data,
         in_uint32_le(in_s, dwReaderLen);
         in_uint8s(in_s, 4); // Referent Identifier
         in_uint32_le(in_s, dwState);
-        dwState = g_ms2pc[dwState % 6];
+        /* dwState is a client-supplied value read into a signed int; use an
+         * unsigned modulo so a high-bit value cannot index g_ms2pc[] negatively. */
+        dwState = g_ms2pc[((unsigned int)dwState) % 6];
         in_uint32_le(in_s, dwProtocol);
         in_uint8a(in_s, attr, 32);
         in_uint32_le(in_s, dwAtrLen);
